Update README.md
I've modified the document behavior of the qualifier based on what we have observed. Technically, if we have artifacts named: `<artifact-name>-<version>.<ex>t` and other artifacts with qualifers named`<artifact-name>-<version>-<qualifier>.<ext>`, if we don't document the qualifiers field, only the artifact without qualifier (plus the version) is retreived, while when adding a list of qualifiers, only artifacts without qualifier plus artificats with one of the specified qualifiers are retreived.